Fix generate key methods and KeyStore bridge
